requirements.txt
accelerate==1.13.0
beautifulsoup4==4.13.4
fastapi==0.115.14
pydantic==2.11.7
python-dotenv==1.2.2
pyyaml==6.0.2
qdrant-client==1.17.1
requests==2.33.0
sentence-transformers==5.3.0
streamlit==1.33.0
torch==2.10.0
transformers==5.3.0
uvicorn==0.35.0
bitsandbytes; platform_system != "Windows"
